# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/coupons": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin Coupon"
                ],
                "summary": "api for admin to see all coupons",
                "operationId": "GetAllCoupons",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully go all the coupons",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get all coupons",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin Coupon"
                ],
                "summary": "api for admin to update the coupon",
                "operationId": "UpdateCoupon",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqEditCoupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully update the coupon",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin Coupon"
                ],
                "summary": "api for admin to add coupon",
                "operationId": "AddCoupon",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCoupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully coupon added",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/carts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can see all productItem that stored in cart",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for get all cart item of user",
                "operationId": "User Cart",
                "responses": {
                    "200": {
                        "description": "successfully got user cart items",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get cart items",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can inrement or drement count of a productItem in cart (min=1)",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for updte productItem count",
                "operationId": "UpdateCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCartCount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully productItem count change on cart"
                    },
                    "400": {
                        "description": "invalid input"
                    },
                    "500": {
                        "description": "can't update the count of product item on cart"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can add a stock in product to user cart",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for add productItem to user cart",
                "operationId": "AddToCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully productItem added to cart"
                    },
                    "400": {
                        "description": "can't add the product item into cart"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can remove a signle productItem full quantity from cart",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for remove a product from cart",
                "operationId": "RemoveFromCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully productItem removed from cart",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "can't remove product item from cart",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/carts/coupons": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "api user to apply on cart on checkout time",
                "operationId": "ApplyCouponToCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqApplyCoupon"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully updated the coupon code",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/coupons": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Coupon"
                ],
                "summary": "api for user to see all coupons",
                "operationId": "GetAllCouponsForUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully go all the coupons",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get all coupons",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.ReqApplyCoupon": {
            "type": "object",
            "required": [
                "coupon_code"
            ],
            "properties": {
                "coupon_code": {
                    "type": "string"
                }
            }
        },
        "req.ReqCart": {
            "type": "object",
            "required": [
                "product_item_id"
            ],
            "properties": {
                "product_item_id": {
                    "type": "integer"
                }
            }
        },
        "req.ReqCartCount": {
            "type": "object",
            "required": [
                "product_item_id"
            ],
            "properties": {
                "count": {
                    "type": "integer",
                    "minimum": 1
                },
                "product_item_id": {
                    "type": "integer"
                }
            }
        },
        "req.ReqCoupon": {
            "type": "object",
            "required": [
                "coupon_name",
                "description",
                "discount_rate",
                "expire_date",
                "image",
                "minimum_cart_price"
            ],
            "properties": {
                "block_status": {
                    "type": "boolean"
                },
                "coupon_name": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 3
                },
                "description": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 6
                },
                "discount_rate": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "expire_date": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "minimum_cart_price": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "req.ReqEditCoupon": {
            "type": "object",
            "required": [
                "coupon_name",
                "description",
                "discount_rate",
                "expire_date",
                "image",
                "minimum_cart_price"
            ],
            "properties": {
                "block_status": {
                    "type": "boolean"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_name": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 3
                },
                "description": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 6
                },
                "discount_rate": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "expire_date": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "minimum_cart_price": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}